refactor(cell): Eliminate `BaseTileShape` Dependency in G2R Loader/Store (#127)
The `BaseTileShape` defined in `include/traits/base.hpp` should be
removed, as the shape of `BaseTile` is not unique and depends on the
instructions used. Various implementations are currently coupled with
it. To address this, this PR introduces the following changes:

- Refine the implementation of the global-to-register loader by
simplifying its structure to reduce overhead. This revision makes it
independent of the `BaseTileShape` defined in `include/traits/base.hpp`.
- Improve the implementation of `Gemm` operator to enhance its
extensibility with the MMA shape.
- Make several minor code cleanups.